+++ src/content-views/utils/__tests__/colors-test.js
